requirements.txt
faiss-cpu==1.12.0
pandas==2.2.3
pyyaml==6.0.1
tqdm==4.66.4
beautifulsoup4==4.12.3
markdownify==0.13.1
lxml
unstructured[local-inference,ocr]>=0.16.19
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.17.0
opencv-python>=4.9.0
PyMuPDF>=1.24.0
python-magic==0.4.27
mammoth>=1.7.0
python-pptx>=0.6.23
openpyxl>=3.1.2
xlrd>=2.0.1
tabulate>=0.9.0
opencv-python-headless>=4.9.0
safetensors==0.6.2
requests==2.32.3
pyarrow
rank_bm25==0.2.2

# Embeddings + reranker (all local)
sentence-transformers
torch>=2.2.0

# (Optional) generation from a local HF model
transformers
accelerate
bitsandbytes>=0.43.1 ; platform_system != "Windows"
openai>=1.30.0
tiktoken>=0.7.0
